SY-4751: move index specs into indexes.spec.ts
